Refactored all CNN CNFUN code to project organization architecture.
- added API calls to LocalDB for configuration MRN refernce. 
- refactored out: constants, enums, initializations, local_odbc, prepare patients, prepare reference, redcap query functions.
