+++ Python/redcap/cnn_and_cnfun_to_redcap.py
@@ -9,6 +9,7 @@
 # History:
 #   -> Summer 2018 - Initial Creation
 #   -> Fall 2018 - Code completion and cleanup
+#   -> Winter 2019 - Code completion and cleanup
 #   -> TBD - Final version
 #
 # Note:
@@ -21,83 +22,24 @@
 #  Imports
 # ----------------------------------------------------------------------------------------------------------------------
 
-from enum import Enum
-from json import dumps
-from operator import itemgetter
-from requests import post  # To install this module (Windows Command Prompt): python -m pip install requests
 from tkinter import *
-
-import pyodbc  # To install this module (Windows Command Prompt): python -m pip install pyodbc
-import sqlite3
-
 
 # ----------------------------------------------------------------------------------------------------------------------
 #  Custom Modules
 # ----------------------------------------------------------------------------------------------------------------------
-
-import globalvars
+from LocalDB.API import load_hospital_record_numbers
+from redcap.prepare_patient import prepare_patient_tables
+from redcap.prepare_reference import prepare_reference_tables
+from redcap.initialization import initialize_data_import_configuration_matrix
+from redcap.query import load_redcap_metadata, send_data_to_redcap
+import logging
 
 # Note: You may tell the script to use the values of the development or production constants simply by commenting out
 #       one of the two lines below.
-import development as environment
 # import production as environment
 
 
-# ----------------------------------------------------------------------------------------------------------------------
-#  Enums
-# ----------------------------------------------------------------------------------------------------------------------
-
-class Project(Enum):
-    admission = 1
-    baby = 2
-    mother = 3
-    patient = 4
-    master = 5
-
-
-class Database(Enum):
-    CNN = 1
-    CNFUN = 2
-
-
-class Field(Enum):
-    Unknown = -1
-    HospitalRecordNumber = 1
-    CaseId = 2
-    BabyId = 3
-    MotherId = 4
-    PatientUI = 5
-    CNNPatientUI = 6
-    PatientId = 7
-    MasterId = 8
-
-
-class DataType(Enum):
-    Unknown = -1
-    Integer = 1
-    String = 2
-
-
-# ----------------------------------------------------------------------------------------------------------------------
-#  Constants
-# ----------------------------------------------------------------------------------------------------------------------
-
-redcap_api_url = environment.REDCAP_API_URL
-
-redcap_token_cnn_admission = environment.REDCAP_TOKEN_CNN_ADMISSION
-redcap_token_cnn_baby = environment.REDCAP_TOKEN_CNN_BABY
-redcap_token_cnn_mother = environment.REDCAP_TOKEN_CNN_MOTHER
-redcap_token_cnn_master = environment.REDCAP_TOKEN_CNN_MASTER
-redcap_token_cnfun_patient = environment.REDCAP_TOKEN_CNFUN_PATIENT
-
-redcap_repeat_instrument_key_name = "redcap_repeat_instrument"
-redcap_repeat_instance_key_name = "redcap_repeat_instance"
-
-redcap_complete_status_suffix = "_complete"
-redcap_complete_status_value = 2
-
-cnn_connection_string = environment.CNN_CONNECTION_STRING
-cnfun_connection_string = environment.CNFUN_CONNECTION_STRING
+logging.basicConfig(stream=sys.stdout, level=logging.INFO)
 
 
 # ----------------------------------------------------------------------------------------------------------------------
@@ -118,7 +60,7 @@ def update_redcap_data():
     label = Label(window, text='')
     label.pack()
 
-    # Initialize the data import configuration matrix.
+    # Load data import configuration matrix.
     label = Label(window, text='Loading Data Import Configuration...')
     label.pack()
     initialize_data_import_configuration_matrix()
@@ -140,16 +82,23 @@ def update_redcap_data():
     label.pack()
 
     # Update Reference Tables.
-    label = Label(window, text='Updating Reference Tables...')
+    label = Label(window, text='Preparing Reference Data Transfer...')
     label.pack()
-    import_reference_tables()
+    prepare_reference_tables()
     label = Label(window, text='Done.')
     label.pack()
 
     # Update Patient Tables.
-    label = Label(window, text='Updating Patient Tables...')
+    label = Label(window, text='Preparing Patient Data Transfer...')
     label.pack()
-    import_patient_tables()
+    prepare_patient_tables()
+    label = Label(window, text='Done.')
+    label.pack()
+
+    # Send data to REDCap.
+    label = Label(window, text='Sending ALL data to REDCap...')
+    label.pack()
+    send_data_to_redcap()
     label = Label(window, text='Done.')
     label.pack()
 
@@ -166,833 +115,30 @@ def update_redcap_data():
     return
 
 
-def load_redcap_metadata():
-    """
-    Get all information about redcap table names and fields.
-    :return: None
-    """
-    globalvars.redcap_metadata = []
+if __name__ == "__main__":
 
-    # For each REDCap project
-    for i in range(1,len(Project) + 1):
+    # ----------------------------------------------------------------------------------------------------------------------
+    #  UI Code
+    # ----------------------------------------------------------------------------------------------------------------------
 
-        # Get all metadata rows.
-        payload = {'token': get_project_token(i), 'format': 'json', 'content': 'metadata'}
-        response = post(redcap_api_url, data=payload)
-        metadata = response.json()
+    # Initialize the tcl/tk interpreter and create the root window.
+    window = Tk()
 
-        # Add each metadata row to local memory.
-        for field in metadata:
-            globalvars.redcap_metadata.append(
-                [field['field_name'], field['field_type'], field['field_label'], field['form_name']])
+    # Adjust size of window.
+    window.geometry("1024x768")
 
-    return
+    # Add a title label to the root window.
+    label = Label(window, text="CNN/CNFUN to REDCap - Data Update")
+    label.pack()
 
+    # Add all buttons to the root window.
+    button = Button(window, text="Update REDCap Data", command=update_redcap_data, height=1, width=25)
+    button.pack()
 
-def load_hospital_record_numbers():
-    """
-    Loads a list of all hospital record numbers for which we want to transfer data to REDCap.
-    :return: None
-    """
-    globalvars.hospital_record_numbers = []
+    # Set window title.
+    window.title("CNN/CNFUN to REDCap - Data Update")
 
-    if environment.USE_LOCAL_HOSPITAL_RECORD_NUMBERS_LIST == 0:
-        # Get the numbers from a dynamic source:
-        db = sqlite3.connect('Local_id_table.sqlite')
-        cursor = db.cursor()
-        cursor.execute('select study_id, MNR from Local_id_table')
+    # Display window.
+    window.mainloop()
 
-        all_rows = cursor.fetchall()
-
-        for row in all_rows:
-            globalvars.hospital_record_numbers.append(row[1])
-
-        cursor.close
-        db.close
-    else:
-        # Get the numbers from a static source:
-        globalvars.hospital_record_numbers = [
-            3143750,
-            3144235,
-            3147383,
-            3149523,
-            3152931,
-            3153280,
-            3154386,
-            3154822,
-            3156430,
-            3160223,
-            3161091,
-            3161116,
-            3161146,
-            3162999,
-            3163000,
-            3163509,
-            3163750,
-            3165201,
-            3165984,
-            3166489,
-            3170659,
-            3171022,
-            3172805,
-            3173436,
-            3174439,
-            3176163,
-            3178972,
-            3181830,
-            3187252,
-            3190535,
-            3191237,
-            3191976,
-            3193639,
-            3202977
-        ]
-    return
-
-
-def initialize_data_import_configuration_matrix():
-    """
-    Initialize the data import configuration matrix.
-    Columns: import_sequence, is_import_enabled, is_reference_table, redcap_project, database_table_name, database,
-             primary_key_name, primary_key_value, authority_on_ids, is_repeatable_instrument, redcap_form_name
-    :return: None
-    """
-    globalvars.data_import_configuration = [
-        [1, 1, 0, 1, 'admission', 1, 1, 1, [2, 3], 0, 'admission'],
-        [2, 1, 0, 1, 'admissionChart', 1, 2, 2, None, 1, 'admissionChart'],
-        [3, 1, 0, 1, 'admissionstate', 1, 2, 2, None, 1, 'admissionstate'],
-        [4, 1, 0, 1, 'Born', 1, 2, 2, None, 1, 'Born'],
-        [5, 1, 0, 1, 'Borndiagnosis', 1, 2, 2, None, 1, 'Borndiagnosis'],
-        [6, 1, 0, 1, 'Caffeine', 1, 2, 2, None, 1, 'Caffeine'],
-        [7, 1, 0, 1, 'Congenitalanomalies', 1, 2, 2, None, 1, 'Congenitalanomalies'],
-        [8, 1, 0, 1, 'Culturestransfusions', 1, 2, 2, None, 1, 'Culturestransfusions'],
-        [9, 1, 0, 1, 'Diagnosisofinfection', 1, 2, 2, None, 1, 'Diagnosisofinfection'],
-        [10, 1, 0, 1, 'Diagnosisprocedures', 1, 2, 2, None, 1, 'Diagnosisprocedures'],
-        [11, 1, 0, 1, 'Discharge', 1, 2, 2, None, 1, 'Discharge'],
-        [12, 1, 0, 1, 'Encephalopathy', 1, 2, 2, None, 1, 'Encephalopathy'],
-        [13, 1, 0, 1, 'Eol', 1, 2, 2, None, 1, 'Eol'],
-        [14, 1, 0, 1, 'Epiqcentrallines', 1, 2, 2, None, 1, 'Epiqcentrallines'],
-        [15, 1, 0, 1, 'Formreviewed', 1, 2, 2, None, 1, 'Formreviewed'],
-        [16, 1, 0, 1, 'Ivh', 1, 2, 2, None, 1, 'Ivh'],
-        [17, 1, 0, 1, 'Nropexams', 1, 2, 2, None, 1, 'Nropexams'],
-        [18, 1, 0, 1, 'Nroptreatments', 1, 2, 2, None, 1, 'Nroptreatments'],
-        [19, 1, 0, 1, 'Ntiss1', 1, 2, 2, None, 1, 'Ntiss1'],
-        [20, 1, 0, 1, 'Ntiss3', 1, 2, 2, None, 1, 'Ntiss3'],
-        [21, 1, 0, 1, 'Otherdiagnosis', 1, 2, 2, None, 1, 'Otherdiagnosis'],
-        [22, 1, 0, 1, 'Pdatreatment', 1, 2, 2, None, 1, 'Pdatreatment'],
-        [23, 1, 0, 1, 'Posbloodcsfcultures', 1, 2, 2, None, 1, 'Posbloodcsfcultures'],
-        [24, 1, 0, 1, 'Postnatalsteroids', 1, 2, 2, None, 1, 'Postnatalsteroids'],
-        [25, 1, 0, 1, 'Posttransfer', 1, 2, 2, None, 1, 'Posttransfer'],
-        [26, 1, 0, 1, 'Rop', 1, 2, 2, None, 1, 'Rop'],
-        [27, 1, 0, 1, 'Screenlog', 1, 2, 2, None, 1, 'Screenlog'],
-        [28, 1, 0, 1, 'Snap', 1, 2, 2, None, 1, 'Snap'],
-        [29, 1, 0, 1, 'Surfactant', 1, 2, 2, None, 1, 'Surfactant'],
-        [30, 1, 0, 1, 'Transport', 1, 2, 2, None, 1, 'Transport'],
-        [31, 1, 0, 1, 'Trips', 1, 2, 2, None, 1, 'Trips'],
-        [32, 1, 0, 1, 'Validate', 1, 2, 2, None, 1, 'Validate'],
-        [33, 1, 0, 2, 'Baby', 1, 3, 3, [4, 5], 0, 'Baby'],
-        [34, 1, 0, 2, 'Bornbaby', 1, 3, 3, None, 1, 'Bornbaby'],
-        [35, 1, 0, 2, 'Epiq', 1, 3, 3, None, 1, 'Epiq'],
-        [36, 1, 0, 2, 'Resuscitation', 1, 3, 3, None, 1, 'Resuscitation'],
-        [37, 1, 0, 2, 'Antenatalinterventions', 1, 3, 3, None, 1, 'Antenatalinterventions'],
-        [38, 1, 0, 3, 'Mother', 1, 4, 4, None, 0, 'Mother'],
-        [39, 1, 0, 3, 'Antenatalintervention', 1, 4, 4, None, 1, 'Antenatalintervention'],
-        [40, 1, 0, 3, 'Bornmother', 1, 4, 4, None, 1, 'Bornmother'],
-        [41, 1, 0, 4, 'Patients', 2, 6, 5, [7], 0, 'Patients'],
-        [42, 1, 0, 4, '18MonthAssessments', 2, 7, 7, None, 1, 'EighteenMonthAssessments'],
-        [43, 1, 0, 4, '36MonthAssessments', 2, 7, 7, None, 1, 'ThirtySixMonthAssessments'],
-        [44, 1, 0, 4, '36MonthStatus', 2, 7, 7, None, 1, 'ThirtySixMonthStatus'],
-        [45, 1, 0, 4, '3YearAuditoryAssessments', 2, 7, 7, None, 1, 'ThreeYearAuditoryAssessments'],
-        [46, 1, 0, 4, '3YearFamilyInfo', 2, 7, 7, None, 1, 'ThreeYearFamilyInfo'],
-        [47, 1, 0, 4, '3YearMedicalHistories', 2, 7, 7, None, 1, 'ThreeYearMedicalHistories'],
-        [48, 1, 0, 4, '3YearPrimaryCaregivers', 2, 7, 7, None, 1, 'ThreeYearPrimaryCaregivers'],
-        [49, 1, 0, 4, '3YearVisionAssessments', 2, 7, 7, None, 1, 'ThreeYearVisionAssessments'],
-        [50, 1, 0, 4, 'AuditoryAssessments', 2, 7, 7, None, 1, 'AuditoryAssessments'],
-        [51, 1, 0, 4, 'FamilyInfo', 2, 7, 7, None, 1, 'FamilyInfo'],
-        [52, 1, 0, 4, 'FollowupReasons', 2, 7, 7, None, 1, 'FollowupReasons'],
-        [53, 1, 0, 4, 'MedicalHistories', 2, 7, 7, None, 1, 'MedicalHistories'],
-        [54, 1, 0, 4, 'PrimaryCaregivers', 2, 7, 7, None, 1, 'PrimaryCaregivers'],
-        [55, 1, 0, 4, 'PsychologicalExaminations', 2, 7, 7, None, 1, 'PsychologicalExaminations'],
-        [56, 1, 0, 4, 'VisionAssessments', 2, 7, 7, None, 1, 'VisionAssessments'],
-        [57, 1, 1, 5, 'mstadmissioncasetype', 1, 8, 8, None, 1, 'mstadmissioncasetype'],
-        [58, 1, 1, 5, 'mstadmissioninfantsex', 1, 8, 8, None, 1, 'mstadmissioninfantsex'],
-        [59, 1, 1, 5, 'mstadmissioninternal', 1, 8, 8, None, 1, 'mstadmissioninternal'],
-        [60, 1, 1, 5, 'mstadmissionpayor', 1, 8, 8, None, 1, 'mstadmissionpayor'],
-        [61, 1, 1, 5, 'mstadmissionstat', 1, 8, 8, None, 1, 'mstadmissionstat'],
-        [62, 1, 1, 5, 'mstborncordbloodteststatus', 1, 8, 8, None, 1, 'mstborncordbloodteststatus'],
-        [63, 1, 1, 5, 'mstborncountries', 1, 8, 8, None, 1, 'mstborncountries'],
-        [64, 1, 1, 5, 'mstborndrugscreenresult', 1, 8, 8, None, 1, 'mstborndrugscreenresult'],
-        [65, 1, 1, 5, 'mstbornhealthcardtype', 1, 8, 8, None, 1, 'mstbornhealthcardtype'],
-        [66, 1, 1, 5, 'mstbornnewbornfeeding', 1, 8, 8, None, 1, 'mstbornnewbornfeeding'],
-        [67, 1, 1, 5, 'mstbornscreeningoffered', 1, 8, 8, None, 1, 'mstbornscreeningoffered'],
-        [68, 1, 1, 5, 'mstbornynu', 1, 8, 8, None, 1, 'mstbornynu'],
-        [69, 1, 1, 5, 'mstbpdgrade', 1, 8, 8, None, 1, 'mstbpdgrade'],
-        [70, 1, 1, 5, 'mstcities', 1, 8, 8, None, 1, 'mstcities'],
-        [71, 1, 1, 5, 'mstcitygroups', 1, 8, 8, None, 1, 'mstcitygroups'],
-        [72, 1, 1, 5, 'mstcnfunhospitallist', 1, 8, 8, None, 1, 'mstcnfunhospitallist'],
-        [73, 1, 1, 5, 'mstculturesculture', 1, 8, 8, None, 1, 'mstculturesculture'],
-        [74, 1, 1, 5, 'mstculturesdiagnosis', 1, 8, 8, None, 1, 'mstculturesdiagnosis'],
-        [75, 1, 1, 5, 'mstculturesorganism', 1, 8, 8, None, 1, 'mstculturesorganism'],
-        [76, 1, 1, 5, 'mstdiagnosiscongenitaltype', 1, 8, 8, None, 1, 'mstdiagnosiscongenitaltype'],
-        [77, 1, 1, 5, 'mstdiagnosisdiagnosistype', 1, 8, 8, None, 1, 'mstdiagnosisdiagnosistype'],
-        [78, 1, 1, 5, 'mstdiagnosishie', 1, 8, 8, None, 1, 'mstdiagnosishie'],
-        [79, 1, 1, 5, 'mstdiagnosisintperforation', 1, 8, 8, None, 1, 'mstdiagnosisintperforation'],
-        [80, 1, 1, 5, 'mstdiagnosisnecstages', 1, 8, 8, None, 1, 'mstdiagnosisnecstages'],
-        [81, 1, 1, 5, 'mstdiagnosispdayes', 1, 8, 8, None, 1, 'mstdiagnosispdayes'],
-        [82, 1, 1, 5, 'mstdiagnosisrds', 1, 8, 8, None, 1, 'mstdiagnosisrds'],
-        [83, 1, 1, 5, 'mstdiagnosisseizures', 1, 8, 8, None, 1, 'mstdiagnosisseizures'],
-        [84, 1, 1, 5, 'mstdiagnosisthrombosislocationarterial', 1, 8, 8, None, 1,
-         'mstdiagnosisthrombosislocationarterial'],
-        [85, 1, 1, 5, 'mstdiagnosisthrombosislocationvenous', 1, 8, 8, None, 1, 'mstdiagnosisthrombosislocationvenous'],
-        [86, 1, 1, 5, 'mstdiagnosisthrombosismethod', 1, 8, 8, None, 1, 'mstdiagnosisthrombosismethod'],
-        [87, 1, 1, 5, 'mstdiagnosisthrombosistreatment', 1, 8, 8, None, 1, 'mstdiagnosisthrombosistreatment'],
-        [88, 1, 1, 5, 'mstdischargedestination', 1, 8, 8, None, 1, 'mstdischargedestination'],
-        [89, 1, 1, 5, 'mstdischargeinpatientarea', 1, 8, 8, None, 1, 'mstdischargeinpatientarea'],
-        [90, 1, 1, 5, 'mstdischargeresult', 1, 8, 8, None, 1, 'mstdischargeresult'],
-        [91, 1, 1, 5, 'mstencephalopathyclinicalstatus', 1, 8, 8, None, 1, 'mstencephalopathyclinicalstatus'],
-        [92, 1, 1, 5, 'mstencephalopathyclinicalstatusatcompletion', 1, 8, 8, None, 1,
-         'mstencephalopathyclinicalstatusatcompletion'],
-        [93, 1, 1, 5, 'mstencephalopathymethod', 1, 8, 8, None, 1, 'mstencephalopathymethod'],
-        [94, 1, 1, 5, 'mstencephalopathyreasonxprematurediscont', 1, 8, 8, None, 1,
-         'mstencephalopathyreasonxprematurediscont'],
-        [95, 1, 1, 5, 'mstencephalopathytargettemp', 1, 8, 8, None, 1, 'mstencephalopathytargettemp'],
-        [96, 1, 1, 5, 'msteol', 1, 8, 8, None, 1, 'msteol'],
-        [97, 1, 1, 5, 'mstepiqchecklistused', 1, 8, 8, None, 1, 'mstepiqchecklistused'],
-        [98, 1, 1, 5, 'mstfacilitytype', 1, 8, 8, None, 1, 'mstfacilitytype'],
-        [99, 1, 1, 5, 'msthospitals', 1, 8, 8, None, 1, 'msthospitals'],
-        [100, 1, 1, 5, 'mstivhresponse', 1, 8, 8, None, 1, 'mstivhresponse'],
-        [101, 1, 1, 5, 'mstivhventricular', 1, 8, 8, None, 1, 'mstivhventricular'],
-        [102, 1, 1, 5, 'mstmotherantecort', 1, 8, 8, None, 1, 'mstmotherantecort'],
-        [103, 1, 1, 5, 'mstmotheranteintervention', 1, 8, 8, None, 1, 'mstmotheranteintervention'],
-        [104, 1, 1, 5, 'mstmotherdeliverytype', 1, 8, 8, None, 1, 'mstmotherdeliverytype'],
-        [105, 1, 1, 5, 'mstmotherdiabetes', 1, 8, 8, None, 1, 'mstmotherdiabetes'],
-        [106, 1, 1, 5, 'mstmotherethnicity', 1, 8, 8, None, 1, 'mstmotherethnicity'],
-        [107, 1, 1, 5, 'mstmotherhypertension', 1, 8, 8, None, 1, 'mstmotherhypertension'],
-        [108, 1, 1, 5, 'mstmotherlabourinit', 1, 8, 8, None, 1, 'mstmotherlabourinit'],
-        [109, 1, 1, 5, 'mstmotherpresentation', 1, 8, 8, None, 1, 'mstmotherpresentation'],
-        [110, 1, 1, 5, 'mstmotherrom', 1, 8, 8, None, 1, 'mstmotherrom'],
-        [111, 1, 1, 5, 'mstmothertotalacgiven', 1, 8, 8, None, 1, 'mstmothertotalacgiven'],
-        [112, 1, 1, 5, 'mstnropcomplications', 1, 8, 8, None, 1, 'mstnropcomplications'],
-        [113, 1, 1, 5, 'mstnropdurations', 1, 8, 8, None, 1, 'mstnropdurations'],
-        [114, 1, 1, 5, 'mstnropfollowup', 1, 8, 8, None, 1, 'mstnropfollowup'],
-        [115, 1, 1, 5, 'mstnropplus', 1, 8, 8, None, 1, 'mstnropplus'],
-        [116, 1, 1, 5, 'mstnropstage', 1, 8, 8, None, 1, 'mstnropstage'],
-        [117, 1, 1, 5, 'mstnroptreatmenteyes', 1, 8, 8, None, 1, 'mstnroptreatmenteyes'],
-        [118, 1, 1, 5, 'mstnroptreatmentlocations', 1, 8, 8, None, 1, 'mstnroptreatmentlocations'],
-        [119, 1, 1, 5, 'mstnroptreatmenttypes', 1, 8, 8, None, 1, 'mstnroptreatmenttypes'],
-        [120, 1, 1, 5, 'mstnropzone', 1, 8, 8, None, 1, 'mstnropzone'],
-        [121, 1, 1, 5, 'mstntissantibiotics', 1, 8, 8, None, 1, 'mstntissantibiotics'],
-        [122, 1, 1, 5, 'mstntisschesttube', 1, 8, 8, None, 1, 'mstntisschesttube'],
-        [123, 1, 1, 5, 'mstntisspacemaker', 1, 8, 8, None, 1, 'mstntisspacemaker'],
-        [124, 1, 1, 5, 'mstntisstracheostomy', 1, 8, 8, None, 1, 'mstntisstracheostomy'],
-        [125, 1, 1, 5, 'mstpatientchart', 1, 8, 8, None, 1, 'mstpatientchart'],
-        [126, 1, 1, 5, 'mstpdatreatment', 1, 8, 8, None, 1, 'mstpdatreatment'],
-        [127, 1, 1, 5, 'mstpostnatalsteroidsindication', 1, 8, 8, None, 1, 'mstpostnatalsteroidsindication'],
-        [128, 1, 1, 5, 'mstpostnatalsteroidsroute', 1, 8, 8, None, 1, 'mstpostnatalsteroidsroute'],
-        [129, 1, 1, 5, 'mstpostnatalsteroidstype', 1, 8, 8, None, 1, 'mstpostnatalsteroidstype'],
-        [130, 1, 1, 5, 'mstpostnatalsteroidstypeindication', 1, 8, 8, None, 1, 'mstpostnatalsteroidstypeindication'],
-        [131, 1, 1, 5, 'mstposttransferdest', 1, 8, 8, None, 1, 'mstposttransferdest'],
-        [132, 1, 1, 5, 'mstprovinces', 1, 8, 8, None, 1, 'mstprovinces'],
-        [133, 1, 1, 5, 'mstresponse', 1, 8, 8, None, 1, 'mstresponse'],
-        [134, 1, 1, 5, 'mstresponsetostimuli', 1, 8, 8, None, 1, 'mstresponsetostimuli'],
-        [135, 1, 1, 5, 'mstrespstatus', 1, 8, 8, None, 1, 'mstrespstatus'],
-        [136, 1, 1, 5, 'mstresuscitationinitialgas', 1, 8, 8, None, 1, 'mstresuscitationinitialgas'],
-        [137, 1, 1, 5, 'mstropplus', 1, 8, 8, None, 1, 'mstropplus'],
-        [138, 1, 1, 5, 'mstropstage', 1, 8, 8, None, 1, 'mstropstage'],
-        [139, 1, 1, 5, 'mstroptreatmenteye', 1, 8, 8, None, 1, 'mstroptreatmenteye'],
-        [140, 1, 1, 5, 'mstroptreatmenttype', 1, 8, 8, None, 1, 'mstroptreatmenttype'],
-        [141, 1, 1, 5, 'mstropzone', 1, 8, 8, None, 1, 'mstropzone'],
-        [142, 1, 1, 5, 'mstsearchgridcolumns', 1, 8, 8, None, 1, 'mstsearchgridcolumns'],
-        [143, 1, 1, 5, 'mstsnapeffectivefio2', 1, 8, 8, None, 1, 'mstsnapeffectivefio2'],
-        [144, 1, 1, 5, 'mstsnapfactor', 1, 8, 8, None, 1, 'mstsnapfactor'],
-        [145, 1, 1, 5, 'mstsnapseizures', 1, 8, 8, None, 1, 'mstsnapseizures'],
-        [146, 1, 1, 5, 'mstsurfactanttype', 1, 8, 8, None, 1, 'mstsurfactanttype'],
-        [147, 1, 1, 5, 'mstsystolicbp', 1, 8, 8, None, 1, 'mstsystolicbp'],
-        [148, 1, 1, 5, 'msttemperature', 1, 8, 8, None, 1, 'msttemperature'],
-        [149, 1, 1, 5, 'mstvegfdrugs', 1, 8, 8, None, 1, 'mstvegfdrugs']
-    ]
-
-    # Sort data import configuration matrix by first column ascending (import sequence).
-    globalvars.data_import_configuration = sorted(globalvars.data_import_configuration, key=itemgetter(0))
-
-    return
-
-
-def import_reference_tables():
-    """
-    Transfers the data of all reference tables to REDCap.
-    :return: None
-    """
-    # For each table in the import configuration matrix
-    for i in range(len(globalvars.data_import_configuration)):
-
-        # If the current table is set to be imported
-        if globalvars.data_import_configuration[i][1]:
-
-            # If the current table is a reference table
-            if globalvars.data_import_configuration[i][2]:
-
-                # Get current table redcap fields.
-                current_table_redcap_fields = get_redcap_fields(globalvars.data_import_configuration[i][10])
-
-                # Get database columns list.
-                database_column_list = get_database_column_names(globalvars.data_import_configuration[i])
-
-                # Get all the data contained in this table.
-                rows = get_data_rows_for_reference_table(globalvars.data_import_configuration[i])
-
-                # Clear REDCap Queue.
-                clear_redcap_queue()
-
-                # For each row in this table
-                for j in range(len(rows)):
-
-                    # Create a blank dictionary.
-                    record_text = {}
-
-                    # Add the ID (always 1 for a reference table)
-                    record_text[get_primary_key_name(globalvars.data_import_configuration[i][6]).lower()] = str(1)
-
-                    # Set repeatable data (if applicable).
-                    if globalvars.data_import_configuration[i][9] == 1:
-                        record_text[redcap_repeat_instrument_key_name] = globalvars.data_import_configuration[i][10].lower()
-                        record_text[redcap_repeat_instance_key_name] = str(j+1)
-
-                    # For each REDCap field in this table
-                    for k in range(len(current_table_redcap_fields)):
-
-                        try:
-
-                            # 0 is for redcap field_label
-                            position_in_database_table = \
-                                database_column_list.index(current_table_redcap_fields[k][0])
-
-                            if str(rows[j][position_in_database_table]) == 'False':
-                                value = '0'
-                            elif str(rows[j][position_in_database_table]) == 'True':
-                                value = '1'
-                            elif str(rows[j][position_in_database_table]) == 'None':
-                                value = ''
-                            else:
-                                value = str(rows[j][position_in_database_table])
-
-                            # 1 is for redcap field_name
-                            record_text[current_table_redcap_fields[k][1]] = str(value)
-
-                        except ValueError:
-
-                            pass
-
-                    # Mark this table entry as 'complete'.
-                    redcap_complete_status_key_name = globalvars.data_import_configuration[i][10].lower() + \
-                                                      redcap_complete_status_suffix
-                    record_text[redcap_complete_status_key_name] = redcap_complete_status_value
-
-                    # Add this item to the REDCap queue.
-                    add_record_to_redcap_queue(record_text)
-
-                    if (j + 1) % environment.NUMBER_OF_RECORDS_PER_BATCH == 0:
-                        # Send records to REDCap
-                        send_data_to_redcap(get_project_token(globalvars.data_import_configuration[i][3]))
-
-                # Send records to REDCap
-                send_data_to_redcap(get_project_token(globalvars.data_import_configuration[i][3]))
-
-    return
-
-
-def import_patient_tables():
-    """
-    Transfers the data of all patient tables to REDCap (only for each hospital record number loaded).
-    :return: None
-    """
-    # For each hospital record number
-    for i in range(len(globalvars.hospital_record_numbers)):
-
-        # Initialize ids.
-        initialize_ids(globalvars.hospital_record_numbers[i])
-
-        # For each table in the import configuration matrix
-        for j in range(len(globalvars.data_import_configuration)):
-
-            # If the current table is set to be imported
-            if globalvars.data_import_configuration[j][1]:
-
-                # If the current table is NOT a reference table
-                if not globalvars.data_import_configuration[j][2]:
-
-                    # Get current table redcap fields.
-                    current_table_redcap_fields = get_redcap_fields(globalvars.data_import_configuration[j][10])
-
-                    # Get database columns list.
-                    database_column_list = get_database_column_names(globalvars.data_import_configuration[j])
-
-                    # Get all data for this patient in this table
-                    rows = get_data_rows_for_patient_table(globalvars.data_import_configuration[j])
-
-                    # If some data was retrieved from the database
-                    if rows is not None:
-
-                        # Clear REDCap Queue.
-                        clear_redcap_queue()
-
-                        # For each row of data retrieved from the database
-                        for k in range(len(rows)):
-
-                            # If this is the first row of data and the current table has authority on any ids
-                            if k == 0 and globalvars.data_import_configuration[j][8] is not None:
-                                for index in range(len(globalvars.data_import_configuration[j][8])):
-
-                                    pk = globalvars.data_import_configuration[j][8][index]
-
-                                    position = database_column_list.index(get_primary_key_name(pk))
-
-                                    if pk == Field.BabyId.value:
-                                        globalvars.BabyId = str(rows[k][position])
-                                    elif pk == Field.CaseId.value:
-                                        globalvars.CaseId = str(rows[k][position])
-                                    elif pk == Field.CNNPatientUI.value:
-                                        globalvars.CNNPatientUI = str(rows[k][position])
-                                    elif pk == Field.HospitalRecordNumber.value:
-                                        globalvars.HospitalRecordNumber = str(rows[k][position])
-                                    elif pk == Field.MotherId.value:
-                                        globalvars.MotherId = str(rows[k][position])
-                                    elif pk == Field.PatientId.value:
-                                        globalvars.PatientId = str(rows[k][position])
-                                    elif pk == Field.PatientUI.value:
-                                        globalvars.PatientUI = str(rows[k][position])
-                                    elif pk == Field.MasterId.value:
-                                        globalvars.MasterId = str(rows[k][position])
-
-                            # Create a blank dictionary.
-                            record_text = {}
-
-                            # Add the ID
-                            pk_for_filter = globalvars.data_import_configuration[j][6]
-                            pk_for_value = globalvars.data_import_configuration[j][7]
-
-                            if pk_for_value == Field.BabyId.value:
-                                record_text[get_primary_key_name(pk_for_filter).lower()] = globalvars.BabyId
-                            elif pk_for_value == Field.CaseId.value:
-                                record_text[get_primary_key_name(pk_for_filter).lower()] = globalvars.CaseId
-                            elif pk_for_value == Field.CNNPatientUI.value:
-                                record_text[Field.PatientId.name.lower()] = globalvars.PatientId
-                            elif pk_for_value == Field.HospitalRecordNumber.value:
-                                # Special Case: We do not save the Hospital Record Number in REDCap.
-                                # Instead, we save the CaseId.'
-                                record_text[Field.CaseId.name.lower()] = globalvars.CaseId
-                            elif pk_for_value == Field.MotherId.value:
-                                record_text[get_primary_key_name(pk_for_filter).lower()] = globalvars.MotherId
-                            elif pk_for_value == Field.PatientId.value:
-                                record_text[get_primary_key_name(pk_for_filter).lower()] = globalvars.PatientId
-                            elif pk_for_value == Field.PatientUI.value:
-                                record_text[Field.PatientId.name.lower()] = globalvars.PatientId
-                            elif pk_for_value == Field.MasterId.value:
-                                record_text[get_primary_key_name(pk_for_filter).lower()] = globalvars.MasterId
-
-                            # Set repeatable data (if applicable).
-                            if globalvars.data_import_configuration[j][9] == 1:
-                                record_text[redcap_repeat_instrument_key_name] = globalvars.data_import_configuration[j][
-                                    10].lower()
-                                record_text[redcap_repeat_instance_key_name] = str(k + 1)
-
-                            # For each REDCap field in this table
-                            for m in range(len(current_table_redcap_fields)):
-
-                                try:
-
-                                    # 0 is for redcap field_label
-                                    position_in_database_table = \
-                                        database_column_list.index(current_table_redcap_fields[m][0])
-
-                                    if str(rows[k][position_in_database_table]) == 'False':
-                                        value = '0'
-                                    elif str(rows[k][position_in_database_table]) == 'True':
-                                        value = '1'
-                                    elif str(rows[k][position_in_database_table]) == 'None':
-                                        value = ''
-                                    else:
-                                        value = str(rows[k][position_in_database_table])
-
-                                    # 1 is for redcap field_name
-                                    record_text[current_table_redcap_fields[m][1]] = str(value)
-
-                                except ValueError:
-
-                                    pass
-
-                            # Mark this table entry as 'complete'.
-                            redcap_complete_status_key_name = globalvars.data_import_configuration[j][10].lower() + \
-                                                              redcap_complete_status_suffix
-                            record_text[redcap_complete_status_key_name] = redcap_complete_status_value
-
-                            # Add this item to the REDCap queue.
-                            add_record_to_redcap_queue(record_text)
-
-                            if (k + 1) % environment.NUMBER_OF_RECORDS_PER_BATCH == 0:
-                                # Send records to REDCap
-                                send_data_to_redcap(get_project_token(globalvars.data_import_configuration[j][3]))
-
-                        # Send records to REDCap
-                        send_data_to_redcap(get_project_token(globalvars.data_import_configuration[j][3]))
-
-    return
-
-
-def add_record_to_redcap_queue(record_text):
-    """
-    Adds a record to the list of records to send to REDCap.
-    :param record_text: Record data
-    :return: None
-    """
-    globalvars.redcap_queue.append(record_text)
-
-    return
-
-
-def get_data_rows_for_reference_table(table_info):
-    """
-    Gets all rows of data for a specific reference table.
-    :param table_info: Table Information
-    :return: List of all the rows obtained from the query.
-    """
-    if table_info is not None:
-        if table_info[4] != '':
-
-            select_statement = 'SELECT * FROM [' + table_info[4] + ']'
-
-            conn = pyodbc.connect(get_connection_string(table_info[5]))
-            odbc_cursor = conn.cursor()
-
-            odbc_cursor.execute(select_statement)
-
-            data = odbc_cursor.fetchall()
-
-            odbc_cursor.close()
-            conn.close
-
-            return data
-
-        else:
-            return None
-    else:
-        return None
-
-
-def get_data_rows_for_patient_table(table_info):
-    """
-    Gets all rows of data for a specific patient table.
-    :param table_info: Table Information
-    :return: List of all the rows obtained from the query.
-    """
-    if table_info is not None:
-        if table_info[4] != '':
-
-            primary_key_filter_name = str(get_primary_key_name(table_info[6]))
-            primary_key_filter_value = str(get_primary_key_value(table_info[7]))
-            primary_key_data_type = get_primary_key_data_type(table_info[7])
-
-            if primary_key_filter_name != '' and primary_key_filter_value != '' and primary_key_filter_value != 'None':
-
-                if primary_key_data_type == DataType.String.value:
-                    select_statement = ("SELECT * FROM [" +
-                                        table_info[4] +
-                                        "] WHERE [" +
-                                        primary_key_filter_name +
-                                        "] = " +
-                                        "'" +
-                                        primary_key_filter_value +
-                                        "'")
-                elif primary_key_data_type == DataType.Integer.value:
-                    select_statement = ("SELECT * FROM [" +
-                                        table_info[4] +
-                                        "] WHERE [" +
-                                        primary_key_filter_name +
-                                        "] = " +
-                                        primary_key_filter_value)
-                else:
-                    select_statement = ''
-
-                if select_statement != '':
-
-                    conn = pyodbc.connect(get_connection_string(table_info[5]))
-                    odbc_cursor = conn.cursor()
-
-                    odbc_cursor.execute(select_statement)
-
-                    data = odbc_cursor.fetchall()
-
-                    odbc_cursor.close()
-                    conn.close
-
-                    return data
-
-                else:
-                    return None
-            else:
-                return None
-        else:
-            return None
-    else:
-        return None
-
-
-def get_redcap_fields(redcap_form_name):
-    """
-    Returns a list of fields contained within a specific REDCap questionnaire (table)
-    :param redcap_form_name: Name of form in REDCap
-    :return: A dictionary (key: Name of field in the database, value: Name of field in REDCap)
-    """
-    redcap_fields = []
-
-    if len(globalvars.redcap_metadata) > 0 and not redcap_form_name == '':
-
-        if redcap_form_name.lower() not in globalvars.redcap_fields:
-
-            # For each REDCap field
-            for i in range(len(globalvars.redcap_metadata)):
-
-                # globalvars.redcap_metadata[i][0] is field_name
-                # globalvars.redcap_metadata[i][2] is field_label
-                # globalvars.redcap_metadata[i][3] is form_name
-
-                if ([globalvars.redcap_metadata[i][2], globalvars.redcap_metadata[i][0]] not in
-                        redcap_fields
-                        and globalvars.redcap_metadata[i][3].lower() == redcap_form_name.lower()):
-                    redcap_fields.append([globalvars.redcap_metadata[i][2],
-                                          globalvars.redcap_metadata[i][0]])
-
-            globalvars.redcap_fields[redcap_form_name.lower()] = redcap_fields
-
-        else:
-
-            redcap_fields = globalvars.redcap_fields[redcap_form_name.lower()]
-
-    return redcap_fields
-
-
-def get_database_column_names(table_info):
-    """
-    Returns a list of fields contained within a database table.
-    :param table_info: Table Information
-    :return: List of fields
-    """
-    if table_info is not None:
-
-        database_columns = ((),)
-
-        if table_info[4].lower() not in globalvars.database_column_names:
-
-            conn = pyodbc.connect(get_connection_string(table_info[5]))
-            odbc_cursor = conn.cursor()
-
-            result = odbc_cursor.execute('SELECT * FROM [' + table_info[4].lower() + '] WHERE 1=0')
-            database_columns = [tuple[0] for tuple in result.description]
-
-            odbc_cursor.close()
-            conn.close
-
-            globalvars.database_column_names[table_info[4].lower()] = database_columns
-
-        else:
-
-            database_columns = globalvars.database_column_names[table_info[4].lower()]
-
-        return database_columns
-
-    else:
-
-        return None
-
-
-def get_project_token(project):
-    """
-    Get Project Token
-    :param project: Project Configuration Number
-    :return: Project Token
-    """
-    if project == Project.admission.value:
-        return redcap_token_cnn_admission
-    elif project == Project.baby.value:
-        return redcap_token_cnn_baby
-    elif project == Project.mother.value:
-        return redcap_token_cnn_mother
-    elif project == Project.master.value:
-        return redcap_token_cnn_master
-    elif project == Project.patient.value:
-        return redcap_token_cnfun_patient
-    else:
-        return ''
-
-
-def get_connection_string(database):
-    """
-    Get Connection String
-    :param database: Database Configuration Number
-    :return: Connection String
-    """
-    if database == Database.CNN.value:
-        return cnn_connection_string
-    elif database == Database.CNFUN.value:
-        return cnfun_connection_string
-    else:
-        return ''
-
-
-def get_primary_key_name(primary_key):
-    """
-    Get Primary Key Name
-    :param primary_key: Primary Key Configuration Number
-    :return: Primary Key Name
-    """
-    if primary_key == Field.BabyId.value:
-        return Field.BabyId.name
-    elif primary_key == Field.CaseId.value:
-        return Field.CaseId.name
-    elif primary_key == Field.CNNPatientUI.value:
-        return Field.CNNPatientUI.name
-    elif primary_key == Field.HospitalRecordNumber.value:
-        return Field.HospitalRecordNumber.name
-    elif primary_key == Field.MotherId.value:
-        return Field.MotherId.name
-    elif primary_key == Field.PatientId.value:
-        return Field.PatientId.name
-    elif primary_key == Field.PatientUI.value:
-        return Field.PatientUI.name
-    elif primary_key == Field.MasterId.value:
-        return Field.MasterId.name
-    else:
-        return Field.Unknown.name
-
-
-def get_primary_key_value(primary_key):
-    """
-    Get Primary Key Value
-    :param primary_key: Primary Key Configuration Number
-    :return: Primary Key Value
-    """
-    if primary_key == Field.BabyId.value:
-        return globalvars.BabyId
-    elif primary_key == Field.CaseId.value:
-        return globalvars.CaseId
-    elif primary_key == Field.CNNPatientUI.value:
-        return globalvars.CNNPatientUI
-    elif primary_key == Field.HospitalRecordNumber.value:
-        return globalvars.HospitalRecordNumber
-    elif primary_key == Field.MotherId.value:
-        return globalvars.MotherId
-    elif primary_key == Field.PatientId.value:
-        return globalvars.PatientId
-    elif primary_key == Field.PatientUI.value:
-        return globalvars.PatientUI
-    elif primary_key == Field.MasterId.value:
-        return globalvars.MasterId
-    else:
-        return -1
-
-
-def get_primary_key_data_type(primary_key):
-    """
-    Get Primary Key Data Type
-    :param primary_key: Primary Key Configuration Number
-    :return: Data Type Configuration Value
-    """
-    if primary_key == Field.BabyId.value:
-        return DataType.Integer.value
-    elif primary_key == Field.CaseId.value:
-        return DataType.String.value
-    elif primary_key == Field.CNNPatientUI.value:
-        return DataType.String.value
-    elif primary_key == Field.HospitalRecordNumber.value:
-        return DataType.String.value
-    elif primary_key == Field.MotherId.value:
-        return DataType.Integer.value
-    elif primary_key == Field.PatientId.value:
-        return DataType.Integer.value
-    elif primary_key == Field.PatientUI.value:
-        return DataType.String.value
-    elif primary_key == Field.MasterId.value:
-        return DataType.Integer.value
-    else:
-        return DataType.Unknown.value
-
-
-def clear_redcap_queue():
-    """
-    Erase all records in the queue.
-    :return: None
-    """
-    globalvars.redcap_queue = []
-
-    return
-
-
-def send_data_to_redcap(project_token):
-    """
-    Sends all records in the queue to REDCap.
-    :param project_token: Token of project to update
-    :return: None
-    """
-    # If there is at least one record in the queue waiting to be sent to REDCap
-    if len(globalvars.redcap_queue) > 0:
-        # If a token was provided
-        if not project_token == '':
-            # Prepare HTTP request
-            payload = {'token': project_token, 'format': 'json', 'content': 'record', 'type': 'flat',
-                       'overwriteBehavior': 'overwrite'}
-            json_string = dumps(globalvars.redcap_queue, separators=(',', ':'))
-            payload['data'] = json_string
-            # Send HTTP request
-            response = post(redcap_api_url, data=payload)
-            # If the request was successful
-            if response.status_code == 200:
-                # Empty the queue.
-                clear_redcap_queue()
-            else:
-                error_message_label = Label(window, text='There was an error in the last API call: ' + response.text)
-                error_message_label.pack()
-
-    return
-
-
-def initialize_ids(hospital_record_number):
-    """
-    Sets the hospital record number and reset all ids related to this hospital record number.
-    :param hospital_record_number: Hospital Record Number
-    :return: None
-    """
-    globalvars.HospitalRecordNumber = hospital_record_number
-    globalvars.CaseId = -1
-    globalvars.BabyId = -1
-    globalvars.MotherId = -1
-    globalvars.PatientUI = -1
-    globalvars.CNNPatientUI = -1
-    globalvars.PatientId = -1
-    globalvars.MasterId = -1
-
-    return
-
-
-# ----------------------------------------------------------------------------------------------------------------------
-#  UI Code
-# ----------------------------------------------------------------------------------------------------------------------
-
-# Initialize the tcl/tk interpreter and create the root window.
-window = Tk()
-
-# Adjust size of window.
-window.geometry("1024x768")
-
-# Add a title label to the root window.
-label = Label(window, text="CNN/CNFUN to REDCap - Data Update")
-label.pack()
-
-# Add all buttons to the root window.
-button = Button(window, text="Update REDCap Data", command=update_redcap_data, height=1, width=25)
-button.pack()
-
-# Set window title.
-window.title("CNN/CNFUN to REDCap - Data Update")
-
-# Display window.
-window.mainloop()
+    update_redcap_data()
