+++ Python/redcap/constants.py
@@ -1,0 +1,17 @@
+from redcap import development as environment
+# ----------------------------------------------------------------------------------------------------------------------
+#  Constants
+# ----------------------------------------------------------------------------------------------------------------------
+
+redcap_api_url = environment.REDCAP_API_URL
+redcap_token_cnn_admission = environment.REDCAP_TOKEN_CNN_ADMISSION
+redcap_token_cnn_baby = environment.REDCAP_TOKEN_CNN_BABY
+redcap_token_cnn_mother = environment.REDCAP_TOKEN_CNN_MOTHER
+redcap_token_cnn_master = environment.REDCAP_TOKEN_CNN_MASTER
+redcap_token_cnfun_patient = environment.REDCAP_TOKEN_CNFUN_PATIENT
+redcap_repeat_instrument_key_name = "redcap_repeat_instrument"
+redcap_repeat_instance_key_name = "redcap_repeat_instance"
+redcap_complete_status_suffix = "_complete"
+redcap_complete_status_value = 2
+cnn_connection_string = environment.CNN_CONNECTION_STRING
+cnfun_connection_string = environment.CNFUN_CONNECTION_STRING
